API: Documenting the task's API
