+++ src/Kyoo.Core/Views/Watch/SubtitleApi.cs
